chore: 식별자 타입 Integer -> Long으로 수정

+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/trim/controller/TrimController.java
@@ -27,7 +27,7 @@ public class TrimController {
 	@Operation(summary = "모델에 따른 트림 목록을 반환합니다")
 	public Response<GetTrimsResponse> getTrims(@RequestBody @Valid GetTrimsRequest getTrimsRequest) {
 
-		Integer modelId = getTrimsRequest.getModelId();
+		Long modelId = getTrimsRequest.getModelId();
 		GetTrimsResponse getTrimsResponse = getTrimsUseCase.execute(modelId);
 		return Response.createSuccessResponse(getTrimsResponse);
 	}
